+++ src/Broadcasting/tests/Middleware/AuthorizationMiddlewareTest.php
@@ -11,6 +11,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Spiral\Broadcasting\AuthorizationStatus;
 use Spiral\Broadcasting\BroadcastInterface;
 use Spiral\Broadcasting\GuardInterface;
 use Spiral\Broadcasting\Middleware\AuthorizationMiddleware;
@@ -71,13 +72,17 @@ final class AuthorizationMiddlewareTest extends TestCase
         $request->shouldReceive('getUri')->once()->andReturn($uri = m::mock(UriInterface::class));
         $uri->shouldReceive('getPath')->once()->andReturn('/auth');
 
-        $broadcast->shouldReceive('authorize')->once()->with($request)->andReturn(true);
+        $broadcast->shouldReceive('authorize')->once()->with($request)
+            ->andReturn(new AuthorizationStatus(
+                true, ['topic_name'], ['foo' => 'bar']
+            ));
+
         $responseFactory->shouldReceive('createResponse')->once()->with(200)->andReturn(m::mock(ResponseInterface::class));
 
         $middleware->process($request, $handler);
     }
 
-    public function testGuardedBroadcastWithNotAuthorizedRequestShouldReturn403Response(): void
+    public function testGuardedBroadcastWithNotAuthorizedRequestShouldReturnForbidResponse(): void
     {
         $middleware = new AuthorizationMiddleware(
             $broadcast = m::mock(BroadcastInterface::class, GuardInterface::class),
@@ -91,9 +96,35 @@ final class AuthorizationMiddlewareTest extends TestCase
         $request->shouldReceive('getUri')->once()->andReturn($uri = m::mock(UriInterface::class));
         $uri->shouldReceive('getPath')->once()->andReturn('/auth');
 
-        $broadcast->shouldReceive('authorize')->once()->with($request)->andReturn(false);
+        $broadcast->shouldReceive('authorize')->once()->with($request)
+            ->andReturn(new AuthorizationStatus(
+                false, ['topic_name'], ['foo' => 'bar']
+            ));
+
         $responseFactory->shouldReceive('createResponse')->once()->with(403)->andReturn(m::mock(ResponseInterface::class));
 
         $middleware->process($request, $handler);
+    }
+
+    public function testGuardedBroadcastWithCustomResponseShouldReturnIt(): void
+    {
+        $middleware = new AuthorizationMiddleware(
+            $broadcast = m::mock(BroadcastInterface::class, GuardInterface::class),
+            $responseFactory = m::mock(ResponseFactoryInterface::class),
+            '/auth',
+        );
+
+        $request = m::mock(ServerRequestInterface::class);
+        $handler = m::mock(RequestHandlerInterface::class);
+
+        $request->shouldReceive('getUri')->once()->andReturn($uri = m::mock(UriInterface::class));
+        $uri->shouldReceive('getPath')->once()->andReturn('/auth');
+
+        $broadcast->shouldReceive('authorize')->once()->with($request)
+            ->andReturn(new AuthorizationStatus(
+                false, ['topic_name'], ['foo' => 'bar'], $response = m::mock(ResponseInterface::class)
+            ));
+
+        $this->assertSame($response, $middleware->process($request, $handler));
     }
 }